Refactor to separate files
